+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 * Fix item bindings not always loading upon login
 * Fix various minor Lua errors when deleting bindings
+* Improve behavior when changing scope of groups
 
 ### Changed